WinGui: Switch the GUI over to using JSON for the queue. (Note, this is not the same format as the CLI and thus cannot be imported. This may happen in the future) Please finish existing queue before installing new version.
