+++ win/CS/HandBrakeWPF/Properties/Resources.resx
@@ -1987,4 +1987,13 @@ Non-Live Options: {date} {time} {creation-date} {creation-time} {quality} {bitra
   <data name="OptionsView_ChoiceOfEncoderHint" xml:space="preserve">
     <value>Choice of encoder will be made available on the 'Video' tab.</value>
   </data>
+  <data name="Queue_RecoverQueueQuestionPlural" xml:space="preserve">
+    <value>HandBrake has detected multiple unfinished queue files. These will be from multiple instances of HandBrake running. Would you like to recover all unfinished jobs?</value>
+  </data>
+  <data name="Queue_RecoverQueueQuestionSingular" xml:space="preserve">
+    <value>HandBrake has detected unfinished items on the queue from the last time the application was launched. Would you like to recover these?</value>
+  </data>
+  <data name="Queue_RecoveryPossible" xml:space="preserve">
+    <value>Queue Recovery Possible</value>
+  </data>
 </root>
