+++ win/CS/HandBrakeWPF/Services/Encode/Model/Models/ChapterMarker.cs
@@ -13,9 +13,9 @@ namespace HandBrakeWPF.Services.Encode.Model.Models
 
     using Caliburn.Micro;
 
-    /// <summary>
-    /// A Movie Chapter
-    /// </summary>
+    using Newtonsoft.Json;
+
+    [JsonObject(MemberSerialization.OptOut)]
     public class ChapterMarker : PropertyChangedBase
     {
         /// <summary>
